backend/order-service/requirements.txt
Flask
pymongo
redis
PyJWT
requests